move ssm functions into makers
