+++ lib/services/parameter-store/make-parameter-store.js
@@ -2,7 +2,7 @@
 
 const Bluebird = require('bluebird');
 const get = require('lodash.get');
-const ssmStore = require('./stores/ssm');
+const { makeSsmStore } = require('./stores/ssm/make-ssm-store');
 const { makeDdbStore } = require('./stores/ddb/make-ddb-store');
 
 const chalk = require('chalk');
@@ -19,7 +19,7 @@ const makeParameterStore = ({ getSettings }) => {
         }
 
         if (get(settings, 'provider.name') === 'ssm') {
-          return ssmStore;
+          return makeSsmStore({});
         }
 
         throw new Error('Unsupported provider specified');
