+++ lib/services/parameter-store/stores/ssm/get-ssm-client.js
@@ -1,0 +1,12 @@
+'use strict';
+
+const AWS = require('aws-sdk');
+
+AWS.config.setPromisesDependency(require('bluebird'));
+AWS.config.update({ region: 'us-east-1' });
+
+const ssm = new AWS.SSM();
+
+module.exports = {
+  getSsmClient: () => ssm
+};
